Day-13/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
streamlit>=1.28.0
numpy>=1.24.0

